+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.0.9 (Unreleased)
 --------------------
 - Fix #139: Missing `parent::init()` in admin controller throws error in HumHub 1.6.beta1
+- Fix #138: Admin only restriction does not include space admin
 
 1.0.8 (May 29, 2020)
 --------------------